requirements.txt
python-telegram-bot==20.3
Faker==20.1.0
pytest==7.4.0
pytest-asyncio==0.21.0
pytest-cov==4.1.0